[TCP] Replace elementwise ops with std elementwise ops.
