+++ lib/Conversion/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_subdirectory(ATenToTCF)
 add_subdirectory(BasicpyToStd)
 add_subdirectory(NumpyToTCF)
+add_subdirectory(TCFToStd)
 add_subdirectory(TCFToTCP)
 
 if(NPCOMP_ENABLE_IREE)
