+++ lib/Conversion/TCFToTCP/TCFToTCP.cpp
@@ -21,112 +21,6 @@
 using namespace mlir;
 using namespace mlir::NPCOMP;
 
-static RankedTensorType getExtentTensorType(Builder &builder) {
-  return RankedTensorType::get({ShapedType::kDynamicSize},
-                               builder.getIndexType());
-}
-
-// Non-templated version of the body of ConvertBinaryElementwise to keep things
-// simple.
-static LogicalResult
-matchAndRewriteBinaryElementwise(Operation *op, PatternRewriter &rewriter) {
-  Value lhs = op->getOperand(0);
-  Value rhs = op->getOperand(1);
-  Location loc = op->getLoc();
-  Value result = op->getResult(0);
-
-  auto lhsType = lhs.getType().dyn_cast<RankedTensorType>();
-  auto rhsType = rhs.getType().dyn_cast<RankedTensorType>();
-  if (!lhsType || !rhsType)
-    return rewriter.notifyMatchFailure(op, "requires ranked tensors");
-
-  Value lhsShape = rewriter.create<shape::ShapeOfOp>(loc, lhs);
-  Value rhsShape = rewriter.create<shape::ShapeOfOp>(loc, rhs);
-
-  // Create the constraints, and the assuming region.
-  Value witness =
-      rewriter.create<shape::CstrBroadcastableOp>(loc, lhsShape, rhsShape);
-  auto assuming = rewriter.create<shape::AssumingOp>(
-      loc, ArrayRef<Type>{result.getType()}, witness);
-
-  // Start building the region body.
-  rewriter.createBlock(&assuming.doRegion());
-  Value broadcastedShape = rewriter.create<shape::BroadcastOp>(
-      loc, getExtentTensorType(rewriter), lhsShape, rhsShape,
-      /*error=*/nullptr);
-
-  // TODO: It's annoying to do the dynamic broadcast above then
-  // do the static transfer function here. Would be nice if they could
-  // somehow be unified.
-  SmallVector<int64_t, 6> broadcastedStaticShape;
-  OpTrait::util::getBroadcastedShape(lhsType.getShape(), rhsType.getShape(),
-                                     broadcastedStaticShape);
-  auto resultType =
-      RankedTensorType::get(broadcastedStaticShape, lhsType.getElementType());
-  Value lhsBroadcasted = rewriter.create<tcp::BroadcastToOp>(
-      loc, resultType, lhs, broadcastedShape);
-  Value rhsBroadcasted = rewriter.create<tcp::BroadcastToOp>(
-      loc, resultType, rhs, broadcastedShape);
-  Value binaryOpResult;
-  if (isa<tcf::AddOp>(op)) {
-    binaryOpResult = rewriter.create<tcp::AddOp>(
-        loc, result.getType(), lhsBroadcasted, rhsBroadcasted);
-  } else if (isa<tcf::MaxOp>(op)) {
-    binaryOpResult = rewriter.create<tcp::MaxOp>(
-        loc, result.getType(), lhsBroadcasted, rhsBroadcasted);
-  } else if (isa<tcf::MulOp>(op)) {
-    binaryOpResult = rewriter.create<tcp::MulOp>(
-        loc, result.getType(), lhsBroadcasted, rhsBroadcasted);
-  } else {
-    op->dump();
-    llvm::report_fatal_error(
-        "unhandled op (see dump above): TCF->TCP binary elementwise");
-  }
-  rewriter.create<shape::AssumingYieldOp>(loc, binaryOpResult);
-
-  // Finally, replace with the results of the shape.assuming
-  rewriter.replaceOp(op, assuming.getResults());
-  return success();
-}
-
-namespace {
-template <typename SourceOp>
-class ConvertBinaryElementwise : public OpRewritePattern<SourceOp> {
-public:
-  using OpRewritePattern<SourceOp>::OpRewritePattern;
-  LogicalResult matchAndRewrite(SourceOp op,
-                                PatternRewriter &rewriter) const override {
-    return matchAndRewriteBinaryElementwise(op, rewriter);
-  }
-};
-} // namespace
-
-static LogicalResult
-matchAndRewriteUnaryElementwise(Operation *op, PatternRewriter &rewriter) {
-  if (isa<tcf::ExpOp>(op)) {
-    rewriter.replaceOpWithNewOp<tcp::ExpOp>(op, op->getOperand(0));
-  } else if (isa<tcf::TanhOp>(op)) {
-    rewriter.replaceOpWithNewOp<tcp::TanhOp>(op, op->getOperand(0));
-  } else {
-    op->dump();
-    llvm::report_fatal_error(
-        "unhandled op (see dump above): TCF->TCP unary elementwise");
-  }
-  return success();
-}
-
-namespace {
-template <typename SourceOp>
-class ConvertUnaryElementwise : public OpRewritePattern<SourceOp> {
-public:
-  using OpRewritePattern<SourceOp>::OpRewritePattern;
-  LogicalResult matchAndRewrite(SourceOp op,
-                                PatternRewriter &rewriter) const override {
-    return matchAndRewriteUnaryElementwise(op, rewriter);
-  }
-};
-} // namespace
-
 namespace {
 class ConvertMatmul : public OpRewritePattern<tcf::MatmulOp> {
 public:
@@ -171,11 +65,6 @@ public:
   FrozenRewritePatternList getPatterns() {
     MLIRContext *context = &getContext();
     OwningRewritePatternList patterns;
-    patterns.insert<ConvertUnaryElementwise<tcf::ExpOp>,
-                    ConvertUnaryElementwise<tcf::TanhOp>>(context);
-    patterns.insert<ConvertBinaryElementwise<tcf::AddOp>,
-                    ConvertBinaryElementwise<tcf::MaxOp>,
-                    ConvertBinaryElementwise<tcf::MulOp>>(context);
     patterns.insert<ConvertMatmul>(context);
     return std::move(patterns);
   }
